+++ backend/PulseGym.DAL/Models/MembershipProgram.cs
@@ -8,7 +8,7 @@ namespace PulseGym.DAL.Models
 
         public required string Name { get; set; }
 
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
         public int Duration { get; set; }
