+++ backend/PulseGym.Logic/DTO/ClientDTO/ClientUpdateDTO.cs
@@ -12,8 +12,8 @@
 
         public string? Goal { get; set; }
 
-        public decimal? InitialWeight { get; set; }
+        public double? InitialWeight { get; set; }
 
-        public decimal? InitialHeight { get; set; }
+        public double? InitialHeight { get; set; }
     }
 }
